# backend/openapi.yaml
openapi: 3.0.3
info:
  title: EliteDial API
  description: Backend API for the EliteDial predictive dialer and call center platform.
  version: 1.0.0
servers:
  - url: /api
    description: Main API (all routes except webhooks)
  - url: /
    description: Root (health check and webhook endpoints)
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
      required: [error]
...
    Call:
      type: object
      properties:
        id:
          type: string
          format: uuid
        provider:
          type: string
          enum: [signalwire, retell, telnyx, mock]
        channel:
          type: string
          enum: [human, ai]
        mode:
          type: string
          enum: [manual, preview, progressive, predictive, ai_outbound, inbound]
        direction:
          type: string
          enum: [inbound, outbound]
        fromNumber:
          type: string
        toNumber:
          type: string
        status:
          type: string
          enum: [initiated, ringing, in-progress, completed, failed, no-answer, busy, voicemail]
        duration:
          type: integer
          description: Duration in seconds
        recordingUrl:
          type: string
          nullable: true
        dispositionId:
          type: string
          nullable: true
        dispositionNote:
          type: string
          nullable: true
        accountId:
          type: string
          nullable: true
        accountName:
          type: string
          nullable: true
        agentId:
          type: string
          format: uuid
          nullable: true
        providerCallId:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        completedAt:
          type: string
          format: date-time
          nullable: true
...
    Voicemail:
      type: object
      properties:
        id:
          type: string
          format: uuid
        fromNumber:
          type: string
        toNumber:
          type: string
        audioUrl:
          type: string
          nullable: true
        transcription:
          type: string
          nullable: true
        duration:
          type: integer
        isRead:
          type: boolean
        assignedToId:
          type: string
          format: uuid
          nullable: true
        assignedTo:
          $ref: '#/components/schemas/UserSummary'
        createdAt:
          type: string
          format: date-time
    UserSummary:
      type: object
      properties:
        id:
          type: string
          format: uuid
        firstName:
          type: string
        lastName:
          type: string
...
security:
  - bearerAuth: []
paths:
...
  /agents/{id}/status:
    patch:
      tags: [Agents]
      summary: Update agent availability status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [status]
              properties:
                status:
                  type: string
                  enum: [available, break, offline, on-call]
      responses:
        '200':
          description: Updated agent status
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                  username:
                    type: string
                  status:
                    type: string
        '403':
          description: Cannot update another agent's status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /calls:
    get:
      tags: [Calls]
      summary: List calls with optional filters
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 25
        - name: direction
          in: query
          schema:
            type: string
            enum: [inbound, outbound]
        - name: status
          in: query
          schema:
            type: string
        - name: agentId
          in: query
          schema:
            type: string
            format: uuid
        - name: startDate
          in: query
          schema:
            type: string
            format: date-time
        - name: endDate
          in: query
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: Paginated call list
          content:
            application/json:
              schema:
                type: object
                properties:
                  calls:
                    type: array
                    items:
                      $ref: '#/components/schemas/Call'
                  total:
                    type: integer
                  page:
                    type: integer
                  limit:
                    type: integer
  /calls/initiate:
    post:
      tags: [Calls]
      summary: Initiate an outbound call
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [toNumber]
              properties:
                toNumber:
                  type: string
                fromNumber:
                  type: string
                mode:
                  type: string
                  enum: [manual, preview, progressive, predictive, ai_outbound]
                accountId:
                  type: string
                accountName:
                  type: string
                campaignId:
                  type: string
                contactId:
                  type: string
                reservationToken:
                  type: string
      responses:
        '200':
          description: Call initiated
          content:
            application/json:
              schema:
                type: object
                properties:
                  callId:
                    type: string
                    format: uuid
                  status:
                    type: string
                  fromNumber:
                    type: string
                  toNumber:
                    type: string
                  provider:
                    type: string
                  mode:
                    type: string
        '400':
          description: Validation error or DNC blocked
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /voicemails:
    get:
      tags: [Voicemails]
      summary: List voicemails (agents see only their own)
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 25
        - name: unreadOnly
          in: query
          schema:
            type: string
            enum: ['true', 'false']
      responses:
        '200':
          description: Paginated voicemails
          content:
            application/json:
              schema:
                type: object
                properties:
                  voicemails:
                    type: array
                    items:
                      $ref: '#/components/schemas/Voicemail'
                  total:
                    type: integer
                  unreadCount:
                    type: integer
                  page:
                    type: integer
                  limit:
                    type: integer
  /voicemails/{id}/read:
    patch:
      tags: [Voicemails]
      summary: Mark a voicemail as read
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Voicemail updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Voicemail'
        '404':
          description: Voicemail not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /voicemails/{id}/assign:
    patch:
      tags: [Voicemails]
      summary: Assign a voicemail to an agent (supervisor+)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [agentId]
              properties:
                agentId:
                  type: string
                  format: uuid
      responses:
        '200':
          description: Voicemail reassigned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Voicemail'

  # ───────────────────────────────────────────────
  # Reports
  # ───────────────────────────────────────────────
...